+++ src/sddc_connector.hpp
@@ -24,7 +24,7 @@ class SddcConnector: public Owrx::Connector {
         virtual int set_center_frequency(double frequency) override;
         virtual int set_sample_rate(double sample_rate) override;
         virtual int set_gain(Owrx::GainSpec* gain) override;
-        virtual int set_ppm(int ppm) override;
+        virtual int set_ppm(double ppm) override;
     private:
         sddc_t* dev;
